requirements.txt
<<<<<<< HEAD
fastapi
uvicorn
passlib
pyjwt
python-jose
python-multipart #form
bcrypt #encripcion


#bearer info segura

=======
fastapi
uvicorn
mysql-connector-python
requests
httpx
pydantic
pydantic[email]

typing
pymongo
motor
spacy
passlib
pyjwt
python-jose
python-multipart
bcrypt
apscheduler


#scraping 
pandas

BeautifulSoup